Update __init__.py
Hi there,

I was using your plugin (which I love btw!) and noticed the hue doesn't change with the H and U commands. I also noticed the transition time was changing. I know very little about code and programming (I actually made this account on git just for this) but I think the error was that the variable transition time was being changed on line 139 and 140 as opposed to hue? I have proposed the change (sorry if I did it wrong). I would love if you could make the change and release it as it is crucial for me to do what I want to do with my octoprint setup! Thank you!

Prash